+++ cypress-tests/cypress/e2e/happyPath/marketplace/commonTestcases/data-source/appWriteHappyPath.cy.skip.js
@@ -28,7 +28,7 @@ describe("Data source AppWrite", () => {
     data.dsName = fake.lastName.toLowerCase().replaceAll("[^A-Za-z]", "");
   });
 
-  it("Should verify elements on appwrite connection form", () => {
+  it.skip("Should verify elements on appwrite connection form", () => {
     const Host = Cypress.env("appwrite_host");
     const ProjectID = Cypress.env("appwrite_projectID");
     const DatabaseID = Cypress.env("appwrite_databaseID");
@@ -100,7 +100,7 @@ describe("Data source AppWrite", () => {
     deleteDatasource(`cypress-${data.dsName}-Appwrite`);
   });
 
-  it("Should verify the functionality of appwrite connection form.", () => {
+  it.skip("Should verify the functionality of appwrite connection form.", () => {
     const Host = Cypress.env("appwrite_host");
     const ProjectID = Cypress.env("appwrite_projectID");
     const DatabaseID = Cypress.env("appwrite_databaseID");
